Fixed error message output during partial cmdline parse, enabling the use of partial cmdline parsing during completion.
Refactored partial argument parsing. Now it uses an OOP approach instead of depending on a method parameter, e.g.
there is a PartialVirtualConsoleArgumentParser class which handles partial cmdline specific workarounds and avoids error messages
getting to the console.

The code got unspaghettified, but the downside is that we need to construct a new argument parser instance for command completion. I don't think
it's expensive and probably we'd do that anyway, since we are planning to construct context dependent argparse confs
based on the current obj state.