Implement a help formatter which renders correctly our special key=value arguments.
Internally we convert key=value arguments to '=key value', so that argparse can handle them as if
they were commandline switches, albeit with an unusual prefix char. Argparse's help will rendere them as such,
so we have to fix that. This commit implements a trivial but limited help message filter which should fix the
help and command usage renderings.