Added an argparser action handler specifically designed to allow us to dynamically declare a group of arguments and later retrieve them as a whole.
Dynamic arguments, such as key=value attributes for the `set` command, would go into the resulting args objects in a python attributed with it's name.
But later during command execution, we might not know the names of those arguments, so it would be impossible to retrieve them as a whole.
This action puts all such arguments into a dict.