Implement a PositionalChoiceCompleter, which completes positional arguments which are choices/enumerations. Shares common positional argument cardinality check with PathCompleter which has been refactored too.
This obsoletes the ObjectTypeCompleter, as the object types are into the argparser syntax itself.
The problem with the ObjectTypeCompleter was that it kept proposing object type names even after the first parameter was set.

Furthermore, the logic which retrieved the list of object types was duplicated in the ObjectTypeCompleter and in the command itself.
Further refactoring of the mk command will filter the types so only those which make sense for the current container will be proposed;
with this patch this single filtering will adapt both help/usage and completion.